pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hurl-orchestra"
version = "0.1.0"
description = "Run hurl test files in dependency order using frontmatter metadata"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities",
    "Environment :: Console",
]
dependencies = [
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "build>=1.0.0",
    "twine>=5.0.0",
]

[project.scripts]
hurl-orchestra = "hurl_orchestra.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hurl_orchestra"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "C90"]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
warn_return_any = true

[[tool.mypy.overrides]]
module = "frontmatter"
ignore_missing_imports = true
